# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DEM",
        "description": "Api documentation",
        "contact": {
            "email": "alessiopiccioli@webmapp.it"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "http://localhost:8000/assets/images/logo_sosec.jpg"
        }
    },
    "paths": {
        "/api/v1/elevation/{lng}/{lat}": {
            "get": {
                "tags": [
                    "Elevation"
                ],
                "summary": "Get Elevation",
                "description": "Retrieve the elevation of a point based on provided longitude and latitude.",
                "operationId": "getElevation",
                "parameters": [
                    {
                        "name": "lng",
                        "in": "path",
                        "description": "Longitude of the point",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "lat",
                        "in": "path",
                        "description": "Latitude of the point",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "elevation": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 0
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Failed to retrieve elevation"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/track": {
            "post": {
                "tags": [
                    "Tech Data"
                ],
                "summary": "Get technical data for a track",
                "operationId": "8fe610cebbc877c15953ad6ec76b91ba",
                "requestBody": {
                    "description": "GeoJSON object that needs to be added to the database",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {},
                            "example": {
                                "type": "Feature",
                                "properties": {
                                    "id": 27060,
                                    "created_at": "2022-09-07T07:47:11.000000Z",
                                    "updated_at": "2024-01-07T22:59:43.000000Z",
                                    "name": {
                                        "it": "119 - Asciano - Scarpa d'Orlando",
                                        "en": "119 - Asciano - Scarpa d'Orlando"
                                    }
                                },
                                "geometry": {
                                    "type": "LineString",
                                    "coordinates": [
                                        [
                                            10.4685531,
                                            43.7494983
                                        ],
                                        [
                                            10.4686648,
                                            43.7494112
                                        ]
                                    ]
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "type": "Feature",
                                    "properties": {
                                        "id": 27060,
                                        "created_at": "2022-09-07T07:47:11.000000Z",
                                        "updated_at": "2024-01-07T22:59:43.000000Z",
                                        "name": {
                                            "it": "119 - Asciano - Scarpa d'Orlando",
                                            "en": "119 - Asciano - Scarpa d'Orlando"
                                        }
                                    },
                                    "geometry": {
                                        "type": "LineString",
                                        "coordinates": [
                                            [
                                                10.4685531,
                                                43.7494983
                                            ],
                                            [
                                                10.4686648,
                                                43.7494112
                                            ]
                                        ]
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        }
    }
}